+++ CHANGELOG.md
@@ -61,6 +61,7 @@ This name should be decided amongst the team before the release.
 - [#1298](https://github.com/topiary/topiary/pull/1298) Update WIT formatter for tree-sitter-wit to v1.4
 - [#1303](https://github.com/topiary/topiary/pull/1303) Add `bin/verify-documented-usage.sh --no-nix-shell` flag
 - [#1306](https://github.com/topiary/topiary/pull/1306) Move topiary-cli configuration logic into its own module
+- [#1311](https://github.com/topiary/topiary/pull/1311) Use `#set!` directive for `injection.language` property
 
 ### Fixed
 - [#1176](https://github.com/topiary/topiary/pull/1176) Increase the stack size to 4MiB in Windows builds.
@@ -77,6 +78,8 @@ This name should be decided amongst the team before the release.
 - [#1254](https://github.com/topiary/topiary/pull/1254) Add Markdown formatting support with dynamic language injections for fenced code blocks (via the `@injection.language` capture).
 - [#1279](https://github.com/topiary/topiary/pull/1279) Added experimental formatting support for the Menhir language.
 - [#1299](https://github.com/topiary/topiary/pull/1299) Add `QuerySource` config fetching to topiary-config.
+- [#1312](https://github.com/topiary/topiary/pull/1312) Added `bin/changelog.nu` to automatically add entry to `CHANGELOG.md`
+- [#1315](https://github.com/topiary/topiary/pull/1315) Add rootcause-backtrace to cli
 
 ### Removed
 - [#1217](https://github.com/topiary/topiary/pull/1217) The `check` alias for the `check-grammar` subcommand, to avoid confusion with `topiary fmt --check`.